+++ app/src/main/java/org/sopt/dosopttemplate/base/BaseActivity.kt
@@ -3,20 +3,26 @@ package org.sopt.dosopttemplate.base
 import android.os.Bundle
 import android.view.MotionEvent
 import android.view.inputmethod.InputMethodManager
+import androidx.activity.OnBackPressedCallback
 import androidx.annotation.LayoutRes
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import androidx.databinding.ViewDataBinding
+import org.sopt.dosopttemplate.util.SnackBar
 
 abstract class BaseActivity<T : ViewDataBinding>(
     @LayoutRes private val layoutRes: Int,
 ) : AppCompatActivity() {
     lateinit var binding: T
 
+    private var backPressedTime = 0L
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = DataBindingUtil.setContentView(this, layoutRes)
         binding.lifecycleOwner = this
+
+        initFinishApp()
     }
 
     override fun dispatchTouchEvent(ev: MotionEvent?): Boolean {
@@ -27,5 +33,23 @@ abstract class BaseActivity<T : ViewDataBinding>(
     private fun hideKeyboard() {
         val inputMethodManager = getSystemService(INPUT_METHOD_SERVICE) as InputMethodManager
         inputMethodManager.hideSoftInputFromWindow(currentFocus?.windowToken, 0)
+    }
+
+    private fun initFinishApp() {
+        onBackPressedDispatcher.addCallback(this, object : OnBackPressedCallback(true) {
+            override fun handleOnBackPressed() {
+                if (System.currentTimeMillis() - backPressedTime >= 2000L) {
+                    backPressedTime = System.currentTimeMillis()
+                    SnackBar.makeSnackBar(binding.root, BACK_MESSAGE)
+                } else {
+                    finishAffinity()
+                }
+            }
+        }
+        )
+    }
+
+    companion object {
+        const val BACK_MESSAGE = "뒤로가기 버튼을 한번 더 누르면 종료됩니다."
     }
 }
